Fixed warnings.

+++ Sources/Kore/IO/Reader.cpp
@@ -29,7 +29,7 @@ u64 Reader::readU64LE(u8* data) {
 #ifdef SYS_LITTLE_ENDIAN
 	return *(u64*)data;
 #else
-	return (data[0] << 0) | (data[1] << 8) | (data[2] << 16) | (data[3] << 24) | (data[4] << 32) | (data[5] << 40) | (data[6] << 48) | (data[7] << 56);
+	return ((u64)data[0] << 0) | ((u64)data[1] << 8) | ((u64)data[2] << 16) | ((u64)data[3] << 24) | ((u64)data[4] << 32) | ((u64)data[5] << 40) | ((u64)data[6] << 48) | ((u64)data[7] << 56);
 #endif
 }
 
@@ -37,7 +37,7 @@ u64 Reader::readU64BE(u8* data) {
 #ifdef SYS_BIG_ENDIAN
 	return *(u64*)data;
 #else
-	return (data[7] << 0) | (data[6] << 8) | (data[5] << 16) | (data[4] << 24) | (data[3] << 32) | (data[2] << 40) | (data[1] << 48) | (data[0] << 56);
+	return ((u64)data[7] << 0) | ((u64)data[6] << 8) | ((u64)data[5] << 16) | ((u64)data[4] << 24) | ((u64)data[3] << 32) | ((u64)data[2] << 40) | ((u64)data[1] << 48) | ((u64)data[0] << 56);
 #endif
 }
 
@@ -45,7 +45,7 @@ s64 Reader::readS64LE(u8* data) {
 #ifdef SYS_LITTLE_ENDIAN
 	return *(s64*)data;
 #else
-	return (data[0] << 0) | (data[1] << 8) | (data[2] << 16) | (data[3] << 24) | (data[4] << 32) | (data[5] << 40) | (data[6] << 48) | (data[7] << 56);
+	return ((s64)data[0] << 0) | ((s64)data[1] << 8) | ((s64)data[2] << 16) | ((s64)data[3] << 24) | ((s64)data[4] << 32) | ((s64)data[5] << 40) | ((s64)data[6] << 48) | ((s64)data[7] << 56);
 #endif
 }
 
@@ -53,7 +53,7 @@ s64 Reader::readS64BE(u8* data) {
 #ifdef SYS_BIG_ENDIAN
 	return *(s64*)data;
 #else
-	return (data[7] << 0) | (data[6] << 8) | (data[5] << 16) | (data[4] << 24) | (data[3] << 32) | (data[2] << 40) | (data[1] << 48) | (data[0] << 56);
+	return ((s64)data[7] << 0) | ((s64)data[6] << 8) | ((s64)data[5] << 16) | ((s64)data[4] << 24) | ((s64)data[3] << 32) | ((s64)data[2] << 40) | ((s64)data[1] << 48) | ((s64)data[0] << 56);
 #endif
 }
